Updated laboratory workflows support
