--- app.docker-compose.yml
@@ -1,25 +1,0 @@
-version: '2.4'
-
-## Port Assignments:
-##  See .env file
-
-## Container debugging: 
-# 1. append the following lines to desired container 
-# 2. boot up the container using `docker-compose up -d`
-# 3 run `docker exec -it <container-name> bash` to start interactive shell
-# 
-#    tty: true
-#    stdin_open: true
-#    entrypoint: bash
-
-## Utility for booting up placeholder page:
-# `docker run --hostname openhim-placeholder --network shared-health-record_sedish --name openhim-placeholder -e MESSAGE=OPENHIM-PLACEHOLDER -e PORT=3000 -d docker.io/sroze/landing-page:latest`
-
-services:
-  shr:
-    image: ghcr.io/i-tech-uw/shared-health-record:main
-    ports:
-      - "3000:3000"
-    restart: on-failure
-    environment:
-      - NODE_ENV=docker
