+++ config/fhir-converter/config.json
@@ -1,0 +1,124 @@
+{
+    "auth": {
+      "secret": "23221o43-71bc-4247-86e4-ea21aaf89c28",
+      "tokenDuration": 5400
+    },
+    "logLevel": "info",
+    "app": {
+      "port": 2019
+    },
+    "mediator": {
+      "api": {
+        "username": "root@openhim.org",
+        "password": "openhim",
+        "apiURL": "https://openhim-core:8080",
+        "trustSelfSigned": true,
+        "urn": "urn:mediator:hl7v2tofhir_translator"
+      }
+    },
+    "mediatorConfig": {
+      "urn": "urn:mediator:hl7v2tofhir_translator",
+      "version": "2.3.4",
+      "name": "HL7v2 to FHIR Translator",
+      "description": "HL7v2 to FHIR Translator",
+      "defaultChannelConfig": [
+        {
+          "methods": [
+            "GET",
+            "POST",
+            "PUT",
+            "PATCH"
+          ],
+          "type": "http",
+          "whitelist": [],
+          "authType": "private",
+          "matchContentTypes": [],
+          "properties": [],
+          "txViewAcl": [],
+          "txViewFullAcl": [],
+          "txRerunAcl": [],
+          "status": "enabled",
+          "rewriteUrls": false,
+          "addAutoRewriteRules": true,
+          "autoRetryEnabled": false,
+          "autoRetryPeriodMinutes": 60,
+          "requestBody": true,
+          "responseBody": true,
+          "name": "HL7v2 Translator API Passthrough",
+          "description": "HL7v2 Translator API Passthrough",
+          "urlPattern": "^/72f/.*$",
+          "routes": [
+            {
+              "type": "http",
+              "status": "enabled",
+              "forwardAuthHeader": false,
+              "name": "Send to Translator ",
+              "secured": false,
+              "host": "fhir-converter",
+              "port": 2019,
+              "path": "",
+              "pathTransform": "s/72f/api/g",
+              "primary": true,
+              "username": "",
+              "password": ""
+            }
+          ],
+          "alerts": [],
+          "rewriteUrlsConfig": [],
+          "priority": 4
+        },
+        {
+          "methods": [
+            "POST"
+          ],
+          "type": "http",
+          "whitelist": [],
+          "authType": "private",
+          "matchContentTypes": [],
+          "properties": [],
+          "txViewAcl": [],
+          "txViewFullAcl": [],
+          "txRerunAcl": [],
+          "status": "enabled",
+          "rewriteUrls": false,
+          "addAutoRewriteRules": true,
+          "autoRetryEnabled": false,
+          "autoRetryPeriodMinutes": 60,
+          "requestBody": true,
+          "responseBody": true,
+          "name": "FHIR --> HL7v2 API",
+          "description": "FHIR to HL7v2 Translation",
+          "urlPattern": "^/f27/.*$",
+          "routes": [
+            {
+              "type": "http",
+              "status": "enabled",
+              "forwardAuthHeader": false,
+              "name": "Send to Translator ",
+              "secured": false,
+              "host": "fhir-converter",
+              "port": 2019,
+              "path": "",
+              "pathTransform": "s/f27/api\/convert\/fhir/g",
+              "primary": true,
+              "username": "",
+              "password": ""
+            }
+          ],
+          "priority": 5
+        }
+      ],
+      "endpoints": [
+        {
+          "name": "HL7v2 to FHIR Translator",
+          "host": "fhir-converter",
+          "port": 2019,
+          "type": "http",
+          "primary": true,
+          "forwardAuthHeader": false,
+          "status": "enabled",
+          "path": "api"
+        }
+      ]
+    }
+  }
