+++ docker-compose.yml
@@ -1,7 +1,4 @@
-version: '2.4'
-
-## Port Assignments:
-##  See .env file
+version: '3.9'
 
 ## Container debugging: 
 # 1. append the following lines to desired container 
@@ -12,29 +9,56 @@ version: '2.4'
 #    stdin_open: true
 #    entrypoint: bash
 
-## Utility for booting up placeholder page:
-# `docker run --hostname openhim-placeholder --network shared-health-record_sedish --name openhim-placeholder -e MESSAGE=OPENHIM-PLACEHOLDER -e PORT=3000 -d docker.io/sroze/landing-page:latest`
-
 services:
+  # shr: Run natively using yarn, or see docker-compose.yml
   shr:
-    hostname: shr
     container_name: shr
-    build: ./
+    hostname: shr
+    build: .
     ports:
-      - "3000:3000"
-    restart: on-failure
+      - 3000:3000
     environment:
       - NODE_ENV=docker
+    volumes:
+      - ./config/config_development.json:/app/config/config_docker.json
+      - ./config/mediator_development.json:/app/config/mediator_docker.json
     networks:
-    - hie
+      - hie
+    profiles: ["shr"]
+    
   shr-fhir:
-    image: "hapiproject/hapi:latest"
     container_name: shr-fhir
     hostname: shr-fhir
+    image: hapiproject/hapi:latest
     restart: unless-stopped
+    volumes:
+      - hapi-data:/data/hapi
+    networks:
+      - hie
+    profiles: ["dev"]
     ports:
-      - "8080:8080"
+      - 8090:8080
 
+  # Kafka Task Runners
+  zookeeper:
+    image: 'bitnami/zookeeper:latest'
+    hostname: zookeeper
+    environment:
+      - ALLOW_ANONYMOUS_LOGIN=yes
+  kafka:
+    image: 'bitnami/kafka:latest'
+    hostname: kafka
+    ports:
+      - '9092:9092'
+    environment:
+      - KAFKA_BROKER_ID=1
+      - KAFKA_CFG_LISTENERS=PLAINTEXT://:9092
+      - KAFKA_CFG_ADVERTISED_LISTENERS=PLAINTEXT://127.0.0.1:9092
+      - KAFKA_CFG_ZOOKEEPER_CONNECT=zookeeper:2181
+      - ALLOW_PLAINTEXT_LISTENER=yes
+      - KAFKA_CFG_AUTO_CREATE_TOPICS_ENABLE=true
+    volumes:
+      - /bitnami/kafka
 networks:
   hie:
     external: true
