+++ src/workflows/__tests__/lab-bw.ts
@@ -1,4 +1,4 @@
-import { LaboratoryWorkflowsBw } from '../lab-bw'
+import { LabWorkflowsBw } from '../LabWorkflowsBw'
 import { R4 } from '@ahryman40k/ts-fhir-types'
 import got from 'got'
 import logger from '../../lib/winston'
@@ -12,7 +12,7 @@ describe('translatePimsCoding', () => {
   it('should translate a given lab test PIMS coding to ciel, loinc, and IPMS', async () => {
     let serviceRequest = <R4.IServiceRequest> (await got.get(IG_URL+"/ServiceRequest-example-bw-pims-service-request-1.json").json())
     
-    let result = await LaboratoryWorkflowsBw.translatePimsCoding(serviceRequest)
+    let result = await LabWorkflowsBw.translatePimsCoding(serviceRequest)
     
     expect(result).toBeDefined
   });
